DT inference: feat: "used core"
I don't totally have a good name for this, but when a set of
observations S doesn't have a decision tree smaller than MAX_SIZE,
you now get a subset S' ⊆ S such that S' on its own doesn't have a
decision tree smaller than MAX_SIZE.

Hopefully S' is small, but there's no guarantee that it's minimal or
anything. Concretely it's the set of observations that were used when
computing lower bounds.

This is analogous to the "UNSAT core" feature of SAT solvers, so maybe
the name "used core" makes sense? IDK