Feat (WIP): CEGAR-ish thing
Suppose for simplicity of explanation we want to prove a lower bound on
DT size for the whole problem. Previously, we would iterate through all
assignments of states to frames ("state assignments"), like so:

```
s = Solver()
while True:
  assignment = s.find_solution()

  prove_bound_on_dt_size(assignment)

  s.add_clause(not(assignment))
```

With the "used core" feature from 5782704, this can become

```
s = Solver()
while True:
  assignment = s.find_solution()

  used_core = prove_bound_on_dt_size(assignment)

  s.add_clause(not(used_core))
```

This is like a "lower bound on DT size" SMT theory! The "used core"
gives us the learned clause. Which is fun. Of course the real code
does't just do this, but actually *finds* the smallest decision tree,
but it's pretty much the same idea.

For the gravity example:
 - Before this feature, it does 48 iterations
 - With this feature, it does 13 iterations